+++ TravellingSalespersonProj/TravellingSalespersonProj/GenericTSP/Graph.cs
@@ -7,17 +7,17 @@ namespace TravellingSalespersonProj
     public class Graph
     {
         // Stores the graph in the format: NodeID, [NodeXPos, NodeYPos]
-        public Dictionary<int, float[]> GraphOfNodes { get; set; }
+        public Dictionary<int, double[]> GraphOfNodes { get; set; }
 
         private readonly RouteEvaluator routeEvaluator;
 
         public Graph()
         {
-            GraphOfNodes = new Dictionary<int, float[]>();
+            GraphOfNodes = new Dictionary<int, double[]>();
             routeEvaluator = new RouteEvaluator();
         }
 
-        public void AddNodeToGraph(int nodeID, float[] nodeCoordinates)
+        public void AddNodeToGraph(int nodeID, double[] nodeCoordinates)
         {
             if (!GraphOfNodes.ContainsKey(nodeID))
             {
@@ -29,7 +29,7 @@ namespace TravellingSalespersonProj
         {
             // Node node 0 (A) had coordinate 0,0 with y traveling downwards with positive values
             // Note this isn't exactly the example
-            float[,] defaultGraph =  new float[,]
+            double[,] defaultGraph =  new double[,]
             {
                 { 0, 0, 0 },
                 { 1, 20, 0 },
@@ -39,7 +39,7 @@ namespace TravellingSalespersonProj
 
             for(int index = 0; index < defaultGraph.GetLength(0); index++)
             {
-                float[] coordinates = new float[] { defaultGraph[index, 1], defaultGraph[index, 2] };
+                double[] coordinates = new double[] { defaultGraph[index, 1], defaultGraph[index, 2] };
                 AddNodeToGraph(Convert.ToInt32(defaultGraph[index, 0]), coordinates);
             }
         }
@@ -47,22 +47,22 @@ namespace TravellingSalespersonProj
         public int GetClosestCity(int currentCity, List<int> visitedCities)
         {
             int closestCity = int.MaxValue;
-            float lowestCostEdge = float.MaxValue;
+            double lowestCostEdge = double.MaxValue;
 
-            float[] currentCityCoordinates = GraphOfNodes[currentCity];
+            double[] currentCityCoordinates = GraphOfNodes[currentCity];
 
-            foreach (KeyValuePair<int, float[]> node in GraphOfNodes)
+            foreach (KeyValuePair<int, double[]> node in GraphOfNodes)
             {
                 if(node.Key == currentCity || visitedCities.Contains(node.Key))
                 {
                     continue;
                 }
 
-                List<float> coordinatesToCompare = new List<float>();
+                List<double> coordinatesToCompare = new List<double>();
                 coordinatesToCompare.AddRange(currentCityCoordinates);
                 coordinatesToCompare.AddRange(node.Value);
 
-                float costOFCurrentEdge = routeEvaluator.CalculateCostOfEdge(coordinatesToCompare.ToArray());
+                double costOFCurrentEdge = routeEvaluator.CalculateCostOfEdge(coordinatesToCompare.ToArray());
 
                 if(costOFCurrentEdge < lowestCostEdge)
                 {
